major restructuring for organization purposes; corrected use of va_arg; compiled successfully
