+++ main.c
@@ -39,6 +39,6 @@ int main(void)
     _printf("Len:[%d]\n", len);
     printf("Len:[%d]\n", len2);
     _printf("Unknown:[%r]\n");
-    printf("Unknown:[%r]\n");
+    /*printf("Unknown:[%r]\n");*/
     return (0);
 }
